+++ apps/web/source.config.ts
@@ -3,16 +3,23 @@ import { defineConfig, defineDocs } from "fumadocs-mdx/config";
 
 export const docs = defineDocs({
   dir: "content/docs",
+  docs: {
+    async: true,
+  },
 });
 
 export default defineConfig({
   mdxOptions: {
     remarkPlugins: [remarkMdxFiles],
+    remarkImageOptions: {
+      onError: "ignore",
+    },
     rehypeCodeOptions: {
       themes: {
         light: "github-light",
         dark: "github-dark",
       },
+      inline: "tailing-curly-colon",
     },
   },
 });