+++ packages/core/src/index.ts
@@ -1,5 +1,0 @@
-// Core package - shared business logic and domain types
-export * from "./constants";
-export * from "./orpc";
-export * from "./rpc-context";
-export * from "./types";